improve key_comparator
Previously this comparator was not implemented for sub-bit-sequences of floats.
Comparsion for entire floats was implemented in a float-wise manner, with special
handling for nan and negative 0. key_comparator is mainly used for radix sorts,
however, and these compare floats in a bitwise manner.

The key_comparator is updated so that it always compares floats in a bitwise manner,
with special handling to make sure that negative values are sorted correctly. In hipCUB
radix sorting algorithms, positive and negative zero are treated the same.

Some extra values are also added in add_special_values to make sure that NaNs with different
values are also sorted as expected.